modified the app.get for movies to make it more modular
